refactor: move functions to @vueuse/shared
