+++ packages/core/index.ts
@@ -7,8 +7,6 @@ export * from './useBrowserLocation'
 export * from './useClipboard'
 export * from './useCounter'
 export * from './useCssVar'
-export * from './useDebounce'
-export * from './useDebounceFn'
 export * from './useDeviceLight'
 export * from './useDeviceMotion'
 export * from './useDeviceOrientation'
@@ -22,8 +20,6 @@ export * from './useFullscreen'
 export * from './useGeolocation'
 export * from './useIdle'
 export * from './useIntersectionObserver'
-export * from './useInterval'
-export * from './useIntervalFn'
 export * from './useLocalStorage'
 export * from './useMediaQuery'
 export * from './useMouse'
@@ -46,10 +42,6 @@ export * from './useSessionStorage'
 export * from './useShare'
 export * from './useSpeechRecognition'
 export * from './useStorage'
-export * from './useThrottle'
-export * from './useThrottleFn'
-export * from './useTimeout'
-export * from './useTimeoutFn'
 export * from './useTitle'
 export * from './useTransition'
 export * from './useWebSocket'
