+++ packages/shared/index.ts
@@ -5,5 +5,13 @@ export * from './makeDestructurable'
 export * from './syncRef'
 export * from './tryOnMounted'
 export * from './tryOnUnmounted'
+export * from './useDebounce'
+export * from './useDebounceFn'
+export * from './useInterval'
+export * from './useIntervalFn'
+export * from './useThrottle'
+export * from './useThrottleFn'
+export * from './useTimeout'
+export * from './useTimeoutFn'
 export * from './utils'
 export * from './when'
