+++ packages/shared/useIntervalFn/index.ts
